+++ lib/store_model/model.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "store_model/json_model_type"
+require "store_model/types"
 
 module StoreModel
   module Model
@@ -10,7 +10,11 @@ module StoreModel
 
       base.extend(Module.new do
         def to_type
-          JsonModelType.new(self)
+          Types::JsonType.new(self)
+        end
+
+        def to_array_type
+          Types::ArrayType.new(self)
         end
       end)
     end