+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+- [PR #5](https://github.com/DmitryTsepelev/store_model/pull/5) Raise error when `#cast` cannot handle the passed instance ([@DmitryTsepelev][])
+- [PR #5](https://github.com/DmitryTsepelev/store_model/pull/5) Add array type generation via Model#to_array_type ([@DmitryTsepelev][])
+
 ## 0.1.2 (2019-03-14)
 
 - `:store_model` validation should not allow nil by default